+++ CHANGELOG.md
@@ -20,8 +20,8 @@ with this project, you should use the latest version.
 * Add reactive model in `reactive.rkt`
 * Reorganize docs to emphasize models
 * Divided source into several multi-collection packages.
-* Add procurement procedure
-* Expand `policy.rkt` to include helpers for `reactive.rkt`
+* Add reactive asset definitions in `modular.rkt`.
+* Expand `policy.rkt` to include helpers for `reactive.rkt`.
 
 ## [1.0] - 2019-11-21
 * Add `#:strict?` keyword argument to compiler